adding invoice search and sales report

+++ Backend/src/PMS/Dto/Medicine/GetMedicineDto.cs
@@ -11,7 +11,7 @@ namespace PMS.Dto.Medicine
         public string BatchCode { get; set; }
         public DateTime MGFDate { get; set; }
         public DateTime ExpiryDate { get; set; }
-        public double Quantity { get; set; }
+        public long Quantity { get; set; }
         public double PricePerUnit { get; set; }
         public long? MedicalCompanyId { get; set; }
         public long UserId { get; set; }
